compute noise coordinates in double, as the reference does
The reference narrows to float at one specific point: after the coordinate is
wrapped, offset and floored, the three fractional parts are cast to float and
everything downstream of that cast — the gradient dot products, the fade curves,
the lerp tree, the octave accumulator — is float. We were doing the coordinate
arithmetic in float too. With origins in [0, 256) that loses about 3e-5 on the
fractional part, which is exactly the drift the oracle was reporting.

Widening the whole kernel to double would have been wrong in the other
direction, moving the lattice stage away from the reference while fixing the
coordinate stage. Only the preamble changes; the octave kernel stays float.

Against the oracle, final_density goes from 493 samples out of tolerance to
none, and its worst error falls from 2.97e-1 to 2.07e-6. The climate roots
improve by two orders of magnitude. The residual is a few float ulp and its
cause is known: the reference sums both octave sets in one accumulator with
normalization folded into each layer, where we keep two accumulators and scale
at the end.

The three pinned samples move with it. They pinned our own output, not the
reference's, and every one of them has moved toward the reference. The legacy
noise pin compared numbers formatted to four decimals, so a 2e-7 shift broke it
on a rounding boundary; it now asserts a tolerance two orders tighter than the
formatting it replaces.

Costs 1.22x on the column benchmark, 1.18 to 1.43 ms. The beta path is
unaffected and its fixtures are byte-identical.

+++ crates/mcrs_minecraft_worldgen/src/noise/normal_noise.rs
@@ -2,7 +2,7 @@ use crate::noise::beta::simplex_octave::SimplexOctaveNoise;
 use crate::noise::octave_perlin_noise::OctavePerlinNoise;
 use mcrs_random::Random;
 
-const INPUT_FACTOR: f32 = 1.0181268882175227;
+const INPUT_FACTOR: f64 = 1.0181268882175227;
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct NormalNoise {
@@ -167,7 +167,7 @@ impl NoiseSampler {
         }
     }
 
-    pub fn get(&self, x: f32, y: f32, z: f32) -> f32 {
+    pub fn get(&self, x: f64, y: f64, z: f64) -> f32 {
         match self {
             Self::Normal(n) => {
                 let x2 = x * INPUT_FACTOR;
@@ -183,8 +183,7 @@ impl NoiseSampler {
             }
             Self::BetaSimplex2d(n) => {
                 n.noise
-                    .sample(x as f64, z as f64, n.scale, n.scale, n.lacunarity, 0.5)
-                    as f32
+                    .sample(x, z, n.scale, n.scale, n.lacunarity, 0.5) as f32
             }
         }
     }
@@ -192,7 +191,7 @@ impl NoiseSampler {
     /// Batch evaluate NoiseSampler at multiple positions (zero heap allocation).
     /// Evaluates both inner OctavePerlinNoise instances in batch, then combines.
     #[cfg(feature = "batch-noise")]
-    pub fn get_batch(&self, positions: &[(f32, f32, f32)], results: &mut [f32]) {
+    pub fn get_batch(&self, positions: &[(f64, f64, f64)], results: &mut [f32]) {
         let n = match self {
             Self::Normal(n) => n,
             _ => {
@@ -208,7 +207,7 @@ impl NoiseSampler {
         debug_assert!(len <= MAX_BATCH);
 
         // Build second-set positions (scaled by INPUT_FACTOR) on stack
-        let mut second_positions = [(0.0f32, 0.0f32, 0.0f32); MAX_BATCH];
+        let mut second_positions = [(0.0f64, 0.0f64, 0.0f64); MAX_BATCH];
         for i in 0..len {
             let (x, y, z) = positions[i];
             second_positions[i] = (x * INPUT_FACTOR, y * INPUT_FACTOR, z * INPUT_FACTOR);
